Refactor of updateserver, please check
Updateserver currently does a git stash and git pull in the case that there are local changes to Git-controlled files.  This causes files to be deleted in the case that there is a Git version that is outdated, which is pretty much all of our files.  This will most likely cause updateserver to take more time, but for the sake of not deleting everyone's TCs, ranks, etc, we should look into this refactor as a serious update to keep in line with our backup process.