Implement room autodelete feature
A necessary feature considering the massive amount of chatrooms we have. This checks to see if more than 10 messages have been sent in the room every hour. If it finds this not to be the case than it begins a countdown of 72 hours, until it either see's 10 more messages or it deletes the room. Rooms can be protected from being deleted by using /protectroom. Furthermore official rooms are not effected by this.