Avatar description tweak
I see loads of people with gif avatars, I set myself one yesterday. I think the problem is with the image hoster: at first my gif would cause a big with the avatar setting command, but when I uploaded it to imgur and tried to set my avatar with the gif I uploaded to imgur it worked perfectly.